Uploading files
The  main source code is the fingerprint file which contains functions for the 4 steps described on Clark and Minella (2016) such as resampling, solve, evaluate confidence region, plot and mores. The quick_start.ipynb  file is a jupyter notebook with basic samples of how to use this codes. The sampledata.xls is a input data file example.